requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.16.2
oauthlib==3.0.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
